Move serializers into separate util module
