+++ broadcast/util/serializers.py
@@ -1,0 +1,31 @@
+import datetime
+import json
+
+
+class DateTimeEncoder(json.JSONEncoder):
+
+    def default(self, obj):
+        if isinstance(obj, datetime.datetime):
+            return obj.isoformat()
+
+        return super(DateTimeEncoder, self).default(obj)
+
+
+class DateTimeDecoder(json.JSONDecoder):
+
+    def __init__(self, *args, **kargs):
+        super(DateTimeDecoder, self).__init__(object_hook=self.object_hook,
+                                              *args,
+                                              **kargs)
+
+    def object_hook(self, obj):
+        if '__type__' not in obj:
+            return obj
+
+        obj_type = obj.pop('__type__')
+        try:
+            return datetime(**obj)
+        except Exception:
+            obj['__type__'] = obj_type
+            return obj
+
